backend/requirements.txt
# Core
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0

# Firebase
firebase-admin==6.6.0

# Redis
redis==5.2.1

# CORS
Flask-CORS==5.0.1
Flask-SocketIO==5.4.1

# Rate Limiting
Flask-Limiter==3.12

# Input Validation
marshmallow==3.25.1

# PDF export
reportlab==4.4.0

# Utilities
uuid6==2024.7.10
requests==2.32.3
boto3==1.39.15

# RAG and vector search
pinecone==7.3.0
sentence-transformers==3.0.1
numpy==2.1.3

# Tests
pytest==8.3.5
pytest-mock==3.14.0
